Annotate for Negation

Annotates EHR text for negation, based on Negex, as simple medical negation lexicon. Tokens are labelled by proximity. The original paper went up to 6 tokens before or after negation.

negex_processor(annotated_text, negex_simp, negex_depth = 6)

Arguments

annotated_text
Dataframe of NLP annotations.

negex_simp
Dataframe of simplified negex.

negex_depth
Maximum distance between word to label and negation term, before or after. Default is 6, as per original paper by Chapman et al.

Value

Dataframe with added negation information.