Process All Documents for One Patient

Performs NLP annotations on all documents using previously established cluster, including NegEx and UMLS CUI tags.

patient_processor_par(
  cl,
  sub_corpus,
  text_format,
  nlp_engine,
  negex_simp,
  umls_selected,
  max_n_grams_length,
  negex_depth,
  single_core_model
)

Arguments

cl
Computing cluster.

sub_corpus
Data frame of text to annotate.

text_format
Text format.

nlp_engine
NLP engine, UDPipe only for now.

negex_simp
Simplifed negex.

umls_selected
Processed UMLS table.

max_n_grams_length
Maximum length of tokens for matching with UMLS concept unique identifiers (CUI's). Shorter values will result in faster processing. If ) is chosen, UMLS CUI tags will not be provided.

negex_depth
Maximum distance between negation item and token to negate. Shorter distances will result in decreased sensitivity but increased specificity for negation.

single_core_model
NLP model in case parallel processing is not used.

Value

NLP annotations dataframe.